+++ Cura/appversion/__init__.py
@@ -1,3 +1,0 @@
-__release_version__ = "2.1.1"
-__build_version__ = "a0"
-__version__ = __release_version__ + __build_version__