+++ Cura/cura.py
@@ -13,6 +13,11 @@ from optparse import OptionParser
 import sys
 import os
 
+os.environ['CURABYDAGO_RELEASE_VERSION'] = '2.1.1'
+os.environ['CURABYDAGO_BUILD_VERSION'] = 'a1'
+os.environ['CURABYDAGO_VERSION'] = os.environ['CURABYDAGO_RELEASE_VERSION'] + os.environ['CURABYDAGO_BUILD_VERSION']
+
+
 # Is it the first run after installation
 if sys.platform.startswith('darwin'):
 	try:
@@ -31,13 +36,12 @@ if sys.platform.startswith('darwin'):
 
 	newinstallfile = os.path.normpath(os.path.join(myResourceBasePath, 'new'))
 	if os.path.isfile(newinstallfile):
-		import Cura.appversion as appv
 		try:
 			os.remove(newinstallfile)
-			current_profile_inifile = os.path.expanduser('~/Library/Application Support/CuraByDagoma/' + appv.__version__ + '/current_profile.ini')
+			current_profile_inifile = os.path.expanduser('~/Library/Application Support/CuraByDagoma/' + os.environ['CURABYDAGO_VERSION'] + '/current_profile.ini')
 			if os.path.isfile(current_profile_inifile):
 				os.remove(current_profile_inifile)
-			preferences_inifile = os.path.expanduser('~/Library/Application Support/CuraByDagoma/' + appv.__version__ + '/preferences.ini')
+			preferences_inifile = os.path.expanduser('~/Library/Application Support/CuraByDagoma/' + os.environ['CURABYDAGO_VERSION'] + '/preferences.ini')
 			if os.path.isfile(preferences_inifile):
 				os.remove(preferences_inifile)
 		except: