+++ Cura/gui/mainWindow.py
@@ -27,8 +27,7 @@ from xml.dom import minidom
 
 class mainWindow(wx.Frame):
 	def __init__(self):
-		import Cura.appversion as appv
-		windowTitle = 'Cura by Dagoma ' + appv.__version__
+		windowTitle = 'Cura by Dagoma ' + os.environ['CURABYDAGO_VERSION']
 		super(mainWindow, self).__init__(None, title=windowTitle, pos=(0, 0), size=wx.DisplaySize())
 
 		wx.EVT_CLOSE(self, self.OnClose)